chore(release): 2.0.0
This has no config changes, but it's a major version bump because if
somebody upgrades, their linting will fail until they add styelint to
their project as a peer dependency. This should be a straightforward
change (and it won't affect production code). The major version bump
will help manage expectations.